requirement.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-groq
langchain_community
langchain-tavily
python-dotenv